Feature-added: Pull To Refresh feature is added. Bug-fixed: Refresh on virgin radio, orientation and fragment incompatibility have been resolved.

+++ Top10/src/com/faraway/top10/types/AbstractMusicList.java
@@ -14,14 +14,11 @@ import java.net.HttpURLConnection;
 import java.net.MalformedURLException;
 import java.net.ProtocolException;
 import java.net.URL;
-import java.net.URLEncoder;
 import java.nio.charset.Charset;
 import java.util.ArrayList;
 import java.util.Locale;
 
 import org.apache.commons.lang.StringUtils;
-import org.jsoup.Jsoup;
-import org.jsoup.nodes.Document;
 
 import android.content.Context;
 
@@ -158,9 +155,12 @@ public abstract class AbstractMusicList {
 		ArrayList<Song> list = readSongList();
 		if (list != null) {
 			for (int i = 0; i < list.size(); i++){
-				File f = new File(list.get(i).fileFullPath);
-				if (f.exists()) {
-					f.delete();
+				String fullPath = list.get(i).fileFullPath;
+				if (fullPath != null) {
+					File f = new File(fullPath);
+					if (f.exists()) {
+						f.delete();
+					}
 				}
 			}
 		}
@@ -173,10 +173,10 @@ public abstract class AbstractMusicList {
 	public abstract String getCacheFileName();
 
 	public abstract String getMusicListName();
-	
+
 	protected String getContentFromURL(String urlString) {
 		String content = new String();
-		
+
 		try {
 			URL url = new URL(urlString);
 			HttpURLConnection urlConnection = (HttpURLConnection) url.openConnection();
@@ -190,7 +190,7 @@ public abstract class AbstractMusicList {
 
 			BufferedReader reader = new BufferedReader(new InputStreamReader(in,Charset.forName("ISO-8859-9")));
 
-			
+
 			String line;
 			while ((line = reader.readLine()) != null){
 				content = content.concat(line);
@@ -198,8 +198,8 @@ public abstract class AbstractMusicList {
 			//Indirilen html dosyasi parse edilerek icindeki sarkilar bulunuyor.
 			//songList = parse(content);
 			in.close();
-			
-						
+
+
 		} catch (MalformedURLException e) {
 			e.printStackTrace();
 		} catch (ProtocolException e) {
@@ -207,7 +207,7 @@ public abstract class AbstractMusicList {
 		} catch (IOException e) {
 			e.printStackTrace();
 		}
-		
+
 		return content;
 	}
